requirements.txt
aiodns==3.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
ccxt==4.4.80
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==44.0.3
distro==1.9.0
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
multidict==6.4.3
nixtla==0.6.6
nixtlats==0.5.2
numpy==1.26.4
openai==1.78.1
orjson==3.10.18
packaging==25.0
pandas==2.2.3
pandas_ta==0.3.14b0
propcache==0.3.1
pycares==4.8.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.3
setuptools==80.4.0
six==1.17.0
sniffio==1.3.1
TA-Lib
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
utilsforecast==0.2.12
websockets==15.0.1
yarl==1.20.0
zstandard==0.23.0
